Fix Equivalent for Ratio
Instead of comparing "component-wise" (which is correct for Equal),
use the mathematical definition: a:b = c:d iff a*d = c*b.

Note that this commit does not handle units correctly in the sense
that for now

  1:2 ~ 3:6                => true (correct)
  1'cm':2'cm' ~ 3'cm':6'm' => true (incorrect)

Once a subsequent commit fixes Multiply and Equivalent for Quantities,
both cases should work.

fixes #1575