Fix Equal, Equivalent for Quantity
If the Quantity.equal method indicates that the units of the supplied
Quantity instances are not comparable by returning null, try to
convert the value of the "right" Quantity to the unit of the "left"
Quantity using the UCUM service.

And similarly for the Quantity.equivalent method with the difference
that the Equivalent returns false if the units are not comparable.

A new method Ratio.fullEquivalent accepts a State and passes it to
EquivalentEvaluator.equivalent so that unit conversion can be used.