Perform unit conversion in the Add, Sub operators for Quantity
AddEvaluator and SubtractEvaluator, if necessary, use the new class
UnitConversionHelper to convert the unit of one of their
operands. This strategy requires the UcumService to be accessible in
the two evaluators which in term means that any evaluator which
performs addition or subtraction now requires the State instance.

A few tests which had to be skipped before this change succeed
now. However, the tests
QtyIvlCollapse_CollapseQuantityUnits[Not]WithinPer now have to be
skipped until comparison operators for Quantity consider unit
conversion.